Split score processing into statistics processor
